# idle-backend/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Max Idle API",
    "version": "1.0.0",
    "description": "Generated API contract for Max Idle backend."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "TimeCurrencyBalances": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "available": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "total",
          "available"
        ]
      },
...
      "Achievement": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "completed": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "icon",
          "completed"
        ]
      },
      "AchievementsResponse": {
        "type": "object",
        "properties": {
          "completedCount": {
            "type": "integer",
            "minimum": 0
          },
          "totalCount": {
            "type": "integer",
            "minimum": 0
          },
          "earningsBonusMultiplier": {
            "type": "number",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "achievements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Achievement"
            }
          }
        },
        "required": [
          "completedCount",
          "totalCount",
          "earningsBonusMultiplier",
          "achievements"
        ]
      },
      "PlayerProfileResponse": {
        "type": "object",
        "properties": {
          "player": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "username": {
                "type": "string"
              },
              "accountAgeSeconds": {
                "type": "integer",
                "minimum": 0
              },
              "currentIdleSeconds": {
                "type": "integer",
                "minimum": 0
              },
              "idleTime": {
                "$ref": "#/components/schemas/TimeCurrencyBalances"
              },
              "realTime": {
                "$ref": "#/components/schemas/TimeCurrencyBalances"
              },
              "timeGems": {
                "$ref": "#/components/schemas/TimeCurrencyBalances"
              },
              "upgradesPurchased": {
                "type": "integer",
                "minimum": 0
              },
              "achievementCount": {
                "type": "integer",
                "minimum": 0
              }
            },
            "required": [
              "id",
              "username",
              "accountAgeSeconds",
              "currentIdleSeconds",
              "idleTime",
              "realTime",
              "timeGems",
              "upgradesPurchased",
              "achievementCount"
            ]
          },
          "meta": {
            "type": "object",
            "properties": {
              "serverTime": {
                "type": "string",
                "format": "date-time"
              }
            },
            "required": [
              "serverTime"
            ]
          }
        },
        "required": [
          "player",
          "meta"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/players/{id}": {
      "get": {
        "tags": [
          "Player"
        ],
        "summary": "Get public player profile",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Player profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerProfileResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid player id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Player not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/achievements": {
      "get": {
        "tags": [
          "Achievements"
        ],
        "summary": "List achievements and completion state",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Achievements",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AchievementsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/achievements/seen": {
      "post": {
        "tags": [
          "Achievements"
        ],
        "summary": "Mark unseen achievements as seen",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Marked as seen"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}